requirements.txt
braintrust-dev>=1.2.0
python-dotenv>=1.0.0
autoevals>=0.3.0 